qa_rag_workspace/requirements.txt
pandas==2.2.2
numpy==1.26.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
chromadb==0.5.4
streamlit==1.37.1
pydantic==2.8.2
openai>=1.35.0
python-dotenv==1.0.1